pack-object & index-pack: add test for --threads warning under NO_PTHREADS
Add a test for the warning that's emitted when --threads or
pack.threads is provided under NO_PTHREADS. This uses the new PTHREADS
prerequisite.

The assertion for !GETTEXT_POISON in the latter test is currently
redundant, since unlike index-pack the pack-objects warnings aren't
i18n'd. However they might be changed to be i18n'd in the future, and
there's no harm in future-proofing the test.

There's an existing bug in the implementation of pack-objects which
this test currently tests for as-is. Details about the bug & the fix
are included in a follow-up change.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>